requirements.txt
torch~=1.7
numpy~=1.19
scipy~=1.6
scikit-learn~=1.0.2
matplotlib
pillow
jupyter
pandas~=1.2
opencv-python~=4.4.0.46
pygifsicle~=1.0.4
imageio~=2.9
openpifpaf~=0.12.13
bs4
lxml~=4.6
jsonlines
tensorboard
hydra-core
cdflib
astropy
seaborn
mlflow
tqdm